refactor: One CraftingStationContainer class
